+++ plugin/resources/assets/js/pages/popup/sections/PlagiarismMatchesSection.vue
@@ -8,9 +8,18 @@
             <v-btn @click="fetchMatches()">Fetch Matches</v-btn>
         </template>
         <div>
+            <v-text-field
+                v-model="search"
+                append-icon="mdi-magnify"
+                label="Search"
+                single-line
+                hide-details
+            ></v-text-field>
             <v-data-table
+                class="center-table"
                 :headers="headers"
-                :items="matches">
+                :items="matches"
+                :search="search">
                 <template v-slot:item.actions="{ item }">
                     <v-row>
                         <plagiarism-match-modal :match="item"></plagiarism-match-modal>
@@ -37,6 +46,7 @@ export default {
 
     data() {
         return {
+            search: "",
             matches: [],
             headers: [
                 {text: 'Matches', align: 'start', value: 'lines_matched'},
@@ -45,7 +55,7 @@ export default {
                 {text: 'Other Uni-ID', value: 'other_uniid'},
                 {text: 'Other Percentage', value: 'other_percentage'},
                 {text: 'Status', value: 'status'},
-                {text: 'Actions', value: 'actions', sortable: false},
+                {text: 'Actions', value: 'actions', sortable: false, filterable: false},
             ]
         }
     },
@@ -63,11 +73,17 @@ export default {
 
             Plagiarism.fetchMatches(this.course.id, this.charon.id, response => {
                 this.matches = response
-
-                this.$emit('matchesFetched', response)
             })
         },
 
     },
 }
 </script>
+<style>
+.center-table table td{
+    vertical-align: middle;
+}
+.center-table table th{
+    vertical-align: middle;
+}
+</style>